requirements.txt
# Core FastAPI and ASGI
fastapi==0.116.1
uvicorn[standard]==0.35.0
starlette==0.47.2
pydantic==2.11.7
pydantic-settings==2.3.0
pydantic-core==2.33.2

# Database and Caching
SQLAlchemy==2.0.41
alembic==1.12.1
psycopg2-binary==2.9.10
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.5
pyotp==2.9.0
email-validator==2.1.0

# HTTP and Networking
httptools==0.6.4
websockets==15.0.1
h11==0.16.0
anyio==4.9.0
sniffio==1.3.1

# File Handling
python-multipart==0.0.20
aiofiles==23.2.1
Pillow==11.3.0

# QR Code and Images
qrcode[pil]==7.4.2

# Payment Processing
stripe==7.11.0

# System Utilities
psutil==5.9.6
PyYAML==6.0.2

# Development and Utilities
python-dotenv==1.1.1
click==8.1.8
watchfiles==1.1.0

# Type Support
typing-extensions==4.14.0
typing-inspection==0.4.1
annotated-types==0.7.0

# Error Handling
exceptiongroup==1.3.0

# Network
idna==3.10

# Performance (Unix/Linux only)
uvloop==0.21.0; sys_platform != "win32"

# Additional Healthcare-specific packages
openpyxl==3.1.2  # Excel file handling
reportlab==4.1.0  # PDF generation
python-dateutil==2.8.2  # Date utilities
pytz==2024.1  # Timezone support
requests==2.32.4  # HTTP requests
jinja2==3.1.3  # Template engine
